# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ISCP API
  version: 0.1.0
  summary: API-first ISCP protocol v2 reference services
servers:
  - url: http://localhost:8080
    description: Relay Reference Service
  - url: http://localhost:8081
    description: Trust Root Reference Service
paths:
...
  /v2/trust/devices/status:
    get:
      summary: Fetch device authorization status
      parameters:
        - name: device_id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: Device status }
  /v2/trust/grants/verify:
    post:
      summary: Verify Trust Grant permission, audience, confirmation, relay, and revocation
      responses:
        "200": { description: Grant valid }
        "403": { description: Grant invalid }
  /v2/trust/grants/status:
    get:
      summary: Fetch Grant status
      parameters:
        - name: grant_id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: Grant status }
...